# QTassessment/QTassessment/bin/Debug/net5.0/Properties/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A mini inventory system.",
    "version": "1.0.0",
    "title": "TMS Inventory",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "matt.luck@landtrafficcontrol.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "tms-inventory.test.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/items": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "Get all items",
        "description": "Endpoint retreives all items in database",
        "operationId": "findItems",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "item"
        ],
        "summary": "Create an item",
        "description": "Endpoint to create an item",
        "operationId": "createItem",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "Created successfully",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "422": {
            "description": "Unprocessable entity"
          }
        }
      }
    },
    "/items/findByItemType": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "Finds Items by ItemType",
        "description": "Multiple ItemType values can be provided with comma separated strings",
        "operationId": "findItemsByItemType",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "ItemType values that need to be considered for filter",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/items/{itemId}": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "Find item by ID",
        "description": "Returns a single item",
        "operationId": "getItemById",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "ID of item to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Item not found"
          }
        }
      },
      "put": {
        "tags": [
          "item"
        ],
        "summary": "Updates a item in the store with form data",
        "description": "",
        "operationId": "updateItemWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "ID of item that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the item",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the item",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "item"
        ],
        "summary": "Deletes a item",
        "description": "",
        "operationId": "deleteItem",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "itemId",
            "in": "path",
            "description": "Item id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ItemType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "ItemType"
      }
    },
    "Item": {
      "type": "object",
      "required": [
        "name",
        "itemTypeId",
        "purchaseOrderNumber"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "itemTypeId": {
          "type": "integer",
          "format": "int64"
        },
        "itemType": {
          "$ref": "#/definitions/ItemType"
        },
        "name": {
          "type": "string",
          "example": "Good Year All Season Deluxe"
        }
      },
      "xml": {
        "name": "Item"
      }
    },
...
  }
}